// src/api/models/HabitKPIsEntry.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HabitKPIs } from './HabitKPIs';
/**
 * One habit's KPIs in a profile-wide read.
 */
export type HabitKPIsEntry = {
    habit_id: number;
    kpis: HabitKPIs;
};